+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 Changes and additions to the library will be listed here.
 
-## 0.1.1 (DEVELOPING)
-- Apply memoization for logger
+## 0.2.0 (DEVELOPING)
+- Dont initialize logger by default anymore.
+- Ensure exaclty once delivery for builtin Kakfa producer.
+- Change the interface of Kafka producer.
+- Emails are packed as msgpack before being put into kafka.
 
 ## 0.1.0
 - Open source the gem.